+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Version History
 v5.20.0
 --------
 
+* MTM1M3TS implementation `<https://github.com/lsst-ts/LOVE-frontend/pull/459>`_
 * Fix scrolling behavior when content overflows on CSCGroup `<https://github.com/lsst-ts/LOVE-frontend/pull/457>`_
 * Fix initial data to avoid errors after creating 1 narrative log `<https://github.com/lsst-ts/LOVE-frontend/pull/456>`_
 * Extend thumbnails query `<https://github.com/lsst-ts/LOVE-frontend/pull/455>`_